global state
