+++ project_pokedex/src/context.js
@@ -1,0 +1,3 @@
+import react from "react";
+
+export const Context = react.createContext()
